Move TokenSigningClients from Integrations to Authentication namespace
